Try fix other platforms

+++ src/hotspot/share/runtime/crac.cpp
@@ -705,8 +705,8 @@ void crac::update_javaTimeNanos_offset() {
 bool crac::read_all(int fd, char *dest, size_t n) {
   size_t rd = 0;
   do {
-    // without the explicit cast it fails the build on Windows
-    ssize_t r = (ssize_t) ::read(fd, dest + rd, n - rd);
+    // using ssize_t fails the build on Windows
+    long r = (long) ::read(fd, dest + rd, n - rd);
     if (r == 0) {
       return false;
     } else if (r < 0) {
